pyproject.toml
[tool.poetry]
name = "knowlege-aggregator"
version = "0.1.0"
description = ""
authors = ["supreet <supreet.shm523@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
langchain-core = "^0.3.28"
duckduckgo-search = "^7.1.0"
bs4 = "^0.0.2"
langgraph = "^0.2.60"
langchain-google-genai = "^2.0.7"
load-dotenv = "^0.1.0"
fastapi = "^0.115.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
